Add modmgr mock
The modmgr.exe is not guaranteed to be existing in all environments and does generate encrypted files.
The mock replicates the cli parameters of the mod manager but creates text files containing what would be inside the mas and rfcmp files.
This aproach makes the generated rfcmp files snapshot testable.